src/currency/api/serializers.py
from rest_framework import serializers
from src.currency.models import Currency


class CurrencySerializer(serializers.ModelSerializer):
    class Meta:
        model = Currency
        fields = [
            'code'
        ]